refactor: refactor folder , files , duplicate code
