--- src/components/ui/kantin/Kantin-Pengiriman-Data/KantinTabel.jsx
@@ -1,144 +1,0 @@
-import {
-  flexRender,
-  getCoreRowModel,
-  getPaginationRowModel,
-  useReactTable,
-} from "@tanstack/react-table";
-
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from "@/components/ui/table";
-import { Button } from "../../button";
-import { useState } from "react";
-import PaginationSelection from "./PaginationSelection";
-import {
-  ChevronLeft,
-  ChevronRight,
-  ChevronsLeft,
-  ChevronsRight,
-} from "lucide-react";
-
-export function KantinTabel({ columns, data }) {
-  const [pagination, setPagination] = useState({
-    pageIndex: 0,
-    pageSize: 10,
-  });
-  const table = useReactTable({
-    data,
-    columns,
-    state: {
-      pagination,
-    },
-    onPaginationChange: setPagination,
-    getCoreRowModel: getCoreRowModel(),
-    getPaginationRowModel: getPaginationRowModel(),
-  });
-
-  return (
-    <div className="overflow-hidden rounded-md">
-      <h1 className=" text-2xl font-bold mb-4">Pengiriman</h1>
-      <Table>
-        <TableHeader>
-          {table.getHeaderGroups().map((headerGroup) => (
-            <TableRow key={headerGroup.id}>
-              {headerGroup.headers.map((header) => {
-                return (
-                  <TableHead key={header.id}>
-                    {header.isPlaceholder
-                      ? null
-                      : flexRender(
-                          header.column.columnDef.header,
-                          header.getContext()
-                        )}
-                  </TableHead>
-                );
-              })}
-            </TableRow>
-          ))}
-        </TableHeader>
-        <TableBody>
-          {table.getRowModel().rows?.length ? (
-            table.getRowModel().rows.map((row) => (
-              <TableRow
-                key={row.id}
-                data-state={row.getIsSelected() && "selected"}
-              >
-                {row.getVisibleCells().map((cell) => (
-                  <TableCell className="py-4" key={cell.id}>
-                    {flexRender(cell.column.columnDef.cell, cell.getContext())}
-                  </TableCell>
-                ))}
-              </TableRow>
-            ))
-          ) : (
-            <TableRow>
-              <TableCell colSpan={columns.length} className="h-24 text-center">
-                No results.
-              </TableCell>
-            </TableRow>
-          )}
-        </TableBody>
-      </Table>
-      <div className="flex items-center justify-between mt-5">
-        <PaginationSelection
-          pagination={pagination}
-          setPagination={setPagination}
-        />
-        <div className="flex gap-4 items-center">
-          <span className="text-sm text-gray-500">
-            Page {pagination.pageIndex + 1} of {table.getPageCount()}
-          </span>
-          <div className="flex items-center justify-end space-x-2 py-4">
-            <Button
-              variant={"outline"}
-              className={"size-9 w-12"}
-              size={"sm"}
-              onClick={() => {
-                table.setPageIndex(0);
-              }}
-              disabled={!table.getCanPreviousPage()}
-            >
-              <ChevronsLeft />
-            </Button>
-          </div>
-          <div className="flex items-center justify-end space-x-2 py-4">
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={() => table.previousPage()}
-              disabled={!table.getCanPreviousPage()}
-            >
-              <ChevronLeft />
-            </Button>
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={() => table.nextPage()}
-              disabled={!table.getCanNextPage()}
-            >
-              <ChevronRight />
-            </Button>
-            <div className="flex items-center justify-end space-x-2 py-4">
-              <Button
-                variant={"outline"}
-                className={"size-9 w-12"}
-                size={"sm"}
-                onClick={() => {
-                  table.setPageIndex(table.getPageCount() - 1);
-                }}
-                disabled={!table.getCanNextPage()}
-              >
-                <ChevronsRight />
-              </Button>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
